packages/dffmpeg-common/pyproject.toml
[project]
name = "dffmpeg-common"
version = "0.1.0"
authors = [
    { name="Bryan Wyatt", email="brwyatt@gmail.com" },
]
description = "Common library functions for dffmpeg applications"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "aiomqtt>=2.5.0, <3.0.0",
    "cryptography>=46.0.0, <47.0.0",
    "httpx>=0.28.0, <1.0.0",
    "pydantic>=2.12.0, <3.0.0",
    "python-ulid[pydantic]>=3.1.0, <4.0.0",
    "pydantic-extra-types[python_ulid]>=2.11.0, <3.0.0",
]

[project.entry-points."dffmpeg.common.crypto"]
fernet = "dffmpeg.common.crypto.fernet:FernetEncryption"

[project.entry-points."dffmpeg.transports.client"]
http_polling = "dffmpeg.common.transports.http_polling:HTTPPollingClientTransport"
mqtt = "dffmpeg.common.transports.mqtt:MQTTClientTransport"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dffmpeg.*"]
namespaces = true

[project.optional-dependencies]
test = [
    "anyio>=4.12.0, <5.0.0",
    "black>=26.1.0, <27.0.0",
    "flake8>=7.3.0, <8.0.0",
    "isort>=7.0.0, <8.0.0",
    "pytest>=9.0.0, <10.0.0",
    "pytest-asyncio>=1.3.0, <2.0.0",
]
dev = [
    "dffmpeg-common[test]",
    "ipython>=9.10.0, <10.0.0",
]
